+++ src/app/app.module.ts
@@ -18,7 +18,7 @@ import { ClientesDetComponent } from './clientes-det/clientes-det.component';
 import { ClientesComponent } from './clientes/clientes.component';
 import { TelaComponent } from './components/tela/tela.component';
 import { HomeComponent } from './home/home.component';
-import { ProdutosComponent } from './produtos/produtos.component';
+import { ProdutosComponent } from './produtos/component/produtos.component';
 
 
 @NgModule({
